Move boltdb session store out of session.go
session.go doesn't directly consume the boltdb session store, it
instead consumes the interface that is provided by it.  Some other
bits and pieces from other source files (realpath(),
secureCookieKeyPair) are only used by the boltdb session.

This change consolidates these pieces into a single place - boltdb.go
- in order to make it easier to re-work some of the surrounding code.
